Add initial version of libtest
This change adds the initial code for libtest, a library providing the
basic infrastructure for creating and running unit tests.
In its design unit tests are made of of test suites which contain
individual test cases. A test case in turn can have multiple test
methods.
On the level of a test case, set up and tear down methods can be
provided that are invoked before and after each test method,
respectively. Once properly organized, a single method invocation can be
used to run all tests and report success/failure.

The library is meant as a simple and lightweight way to get a unit
testing facility into a program. Creating a true library (that is, a
static archive or a dynamic object) is out of scope and a header only
solution is what is being aimed for. Plus, most methods are so small
that inlining everything might not cause any bloat at all but rather
create a smaller end result.