Do not trip over empty messages
The ASSERT macro does not pass in a message to display to the framework
(as oppossed to ASSERTM). The method handling the assertion in the
DefaultResult will, thus, receive an empty, i.e., NULL, message which we
did not anticipate.
This change fixes the problem by providing a proper check.