Add sample file illustrating usage
Writing a unit test should be straight forward. Usually, the user should
not be required to have to lookup the documentation for simple tasks.
Instead, a sample file illustrating basic usage patterns and best
practices helps in lowering the steepness of the learning curve quite a
bit.
This change adds such a sample file useful to use as a base to work with
or as quick reference.