Introduce concept of a printer for DefaultResult class
In its initial version, the DefaultResult used an std::iostream to
convey any messages or results to the user. This approach is not very
flexible when it comes to different environments. For instance, if no
C++ standard library exists the DefaultResult class could not be used at
all and its functionality would have to be reimplemented.
This change introduces an indirection in the form of a printer object
that can be passed to DefaultResult objects and will be used for
printing any data.