Add virtual destructor to base classes provided by the framework
In general, when creating class hierarchies and interacting with
objects of different types, it is good practise for the very base
classes to have a virtual destructor. Not doing so becomes a problem
once one tries to delete derived objects through a pointer to a base
class.
Arguably, no such scenarios are envisioned as of now but then clients
should also not be restricted in their usage of this library.
This change adds virtual destructors to all base classes to avoid this
potential problem.