Use new C++ features to remove certain compiler generated functions
By now most compilers support C++11 and it is considered ready for prime
time. The new standard provides relevant advantages in all sorts of
scenarios.
As part of the general switch to employing it throughout the entire code
base, this change adjusts libtest. In particular this patch makes use of
the 'delete' keyword to remove unwanted compiler-generated methods (such
as the copy constructor and assignment operator).