Use libutil's assertion backend
libutil contains an assertion implementation that provides some more
sophisticated checking functionality. For libtest it makes sense to make
use of this backend in order to reduce code duplication and to stick to
a common base.
This change adjusts the assertion macros in libtest to make use of
libutil's macro backend. To comply with the implementation, we have to
split our single assert method into two: one that merely informs the
framework that an assertion is being checked (which is useful for
gathering statistics) and another one that acknowledges (and reports) a
failed assertion.