Introduce TESTASSERTOP
With the usage of the new assertion backend provided by libutil, we can
also support another type of assertion that includes slightly more
information in the reported error when a failure occurs: TESTASSERTOP
checks the result of a binary operation. If the assertion fails (i.e.,
the result is not true), the values of both operands are included in the
reported error message (along with the operation, that is).