Include AssertImpl.hpp header instead of Assert.hpp
The first version of assertion support in libutil offered a single
header providing the full functionality: Assert.hpp.
This turned out to not work properly when multiple types of assertions
are based on top of the implementation therein. As such, the header was
split into Assert.hpp and AssertImpl.hpp -- the latter providing the
actual implementation and the former being merely the one offering a
default implementation.
Since libtest implements its own assertions, we should now include
AssertImpl.hpp instead of Assert.hpp.