Add TESTTHROWS set of macros
This change adds a couple of macros that can be used to assert that an
expression raises an exception. The set of macros comprises those that
check for a specific type of exception and those for any. It also
includes versions that convey an error message to the user in case of
assertion failure.